requirements-local.txt
# Local development and testing requirements
# Includes everything needed for local dev, testing, linting, and CI

# Include production dependencies
-r requirements.txt

# Dependencies pre-installed in EMR (needed locally)
pyspark>=3.3.0,<4.0.0
boto3>=1.26.0
botocore>=1.29.0
pandas>=1.5.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
testcontainers[postgres]>=4.0.0
moto[s3]>=5.0.0

# Code formatting and linting
black>=25.11.0,<26.0.0
isort>=5.12.0
flake8>=6.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# Security scanning
bandit>=1.7.0
safety>=2.3.0

# Geospatial processing
apache-sedona>=1.5.0
